+++ packages/omni-auth/src/types.ts
@@ -7,7 +7,6 @@ export interface PublicUser {
   id: string;
   name: string;
   email: string;
-  emailVerified: boolean;
   image: string | null;
   createdAt: Date;
   updatedAt: Date;